+++ noodles-gff/src/async/io/reader.rs
@@ -3,7 +3,7 @@ mod line;
 use futures::{stream, Stream, TryStreamExt};
 use tokio::io::{self, AsyncBufRead, AsyncBufReadExt};
 
-use crate::{DirectiveBuf, Line, LineBuf, RecordBuf};
+use crate::{directive_buf::key, DirectiveBuf, Line, LineBuf, RecordBuf};
 
 /// An async GFF reader.
 pub struct Reader<R> {
@@ -178,7 +178,12 @@ where
         Box::pin(stream::try_unfold(self.line_bufs(), |mut lines| async {
             loop {
                 match lines.try_next().await? {
-                    None | Some(LineBuf::Directive(DirectiveBuf::StartOfFasta)) => return Ok(None),
+                    None => return Ok(None),
+                    Some(LineBuf::Directive(DirectiveBuf::Other(key, _)))
+                        if key == key::START_OF_FASTA =>
+                    {
+                        return Ok(None)
+                    }
                     Some(LineBuf::Record(record)) => return Ok(Some((record, lines))),
                     _ => {}
                 }