+++ noodles-gff/src/io/reader/record_bufs.rs
@@ -1,6 +1,6 @@
 use std::io::{self, BufRead};
 
-use crate::{DirectiveBuf, LineBuf, RecordBuf};
+use crate::{directive_buf::key, DirectiveBuf, LineBuf, RecordBuf};
 
 use super::LineBufs;
 
@@ -33,7 +33,11 @@ where
         loop {
             match self.lines.next()? {
                 Ok(line) => match line {
-                    LineBuf::Directive(DirectiveBuf::StartOfFasta) => return None,
+                    LineBuf::Directive(DirectiveBuf::Other(key, _))
+                        if key == key::START_OF_FASTA =>
+                    {
+                        return None
+                    }
                     LineBuf::Record(r) => return Some(Ok(r)),
                     _ => {}
                 },